% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2z-package.R
\docType{package}
\name{c2z-package}
\alias{c2z}
\alias{c2z-package}
\title{c2z: A Reference Manager}
\description{
'Cristin' to 'Zotero' ('c2z') aims at obtaining total dominion over 'Cristin' ('Current Research Information SysTem in Norway') and 'Zotero'. The package enables manipulating 'Zotero' libraries using 'R'. Import references from 'Cristin', 'Regjeringen', 'CRAN', 'ISBN' ('Alma', 'LoC'), and 'DOI' ('CrossRef', 'DataCite') to a 'Zotero' library. Add, edit, copy, or delete items, including attachments and collections, and export references to 'BibLaTeX' (and other formats).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/oeysan/c2z/}
  \item \url{https://oeysan.github.io/c2z/}
  \item Report bugs at \url{https://github.com/oeysan/c2z/issues/}
}

}
\author{
\strong{Maintainer}: Øystein Olav Skaar \email{oeysan@gmail.com} (\href{https://orcid.org/0000-0001-5605-3492}{ORCID}) [copyright holder]

}
\keyword{internal}
