% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2z.R
\docType{data}
\name{cristin.types}
\alias{cristin.types}
\title{List with empty cristin-items}
\format{
A list with 70 tibbles with zero rows and various columns
}
\usage{
cristin.types
}
\description{
Each tibble in the list represents a cristin-item
}
\details{
Used to create Zotero-items from list of metadata
}
\keyword{internal}
