% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cristin_wrangler.R
\name{CristinWrangler}
\alias{CristinWrangler}
\title{Wrangle Cristin metadata into Zotero-type format}
\usage{
CristinWrangler(data, meta = list(), use.identifiers = TRUE)
}
\arguments{
\item{data}{Tibble containing metadata from Cristin}

\item{meta}{A list for storing data elements, Default: list()}

\item{use.identifiers}{Use if ISBN/DOI identifiers if enabled, Default: TRUE}
}
\value{
A Zotero-type tibble.
}
\description{
Converts Cristin metadata into a Zotero-style tibble.
}
\details{
Please see \href{https://oeysan.github.io/c2z/}{https://oeysan.github.io/c2z/}
}
