% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi_datacite.R
\name{DoiDatacite}
\alias{DoiDatacite}
\title{Wrangle DataCite metadata into Zotero-type format}
\usage{
DoiDatacite(data, meta)
}
\arguments{
\item{data}{XML data from DataCite containing metadata}

\item{meta}{A list collecting all metadata used to create , Default: list()}
}
\value{
A Zotero-type matrix (tibble)
}
\description{
Query DataCite by DOI and fetch metadata
}
