% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2z.R
\docType{data}
\name{zotero.types}
\alias{zotero.types}
\title{List with empty zotero-items}
\format{
A list with 36 tibbles with zero rows and various columns
}
\usage{
zotero.types
}
\description{
Each tibble in the list represents a zotero-item
}
\details{
Used to create Zotero-items from list of metadata
}
\keyword{internal}
