% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero_check.R
\name{ZoteroCheck}
\alias{ZoteroCheck}
\title{Check Zotero library for duplicates}
\usage{
ZoteroCheck(
  data,
  id,
  id.type,
  created,
  last.modified,
  zotero = NULL,
  silent = FALSE,
  log = list()
)
}
\arguments{
\item{data}{Tibble containing metadata (from Cristin)}

\item{id}{column containing identifier (e.g., cristin_result_id)}

\item{id.type}{Type of metadata as found in the Zotero extra field ("e.g., Cristin)}

\item{created}{column containing creation date in UNIX timestamp format (e.g., created)}

\item{last.modified}{column containing modifcation date in UNIX timestamp format (e.g., last_modified)}

\item{zotero}{A list with information on the specified Zotero library (e.g.,
id, API key, collections, and items), Default: NULL}

\item{silent}{c2z is noisy, tell it to be quiet, Default: FALSE}

\item{log}{A list for storing log elements, Default: list()}
}
\value{
Returns non-duplicated data in as Zotero-type matrix (tibble)
}
\description{
Remove references that are not modified since last added to Zotero
}
\seealso{
\code{\link[dplyr]{arrange}},
 \code{\link[dplyr]{coalesce}},
 \code{\link[dplyr]{bind_rows}},
 \code{\link[dplyr]{filter}}
}
