% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero_doi.R
\name{ZoteroDoi}
\alias{ZoteroDoi}
\title{Use DOI to acquire metadata}
\usage{
ZoteroDoi(doi, meta = list())
}
\arguments{
\item{doi}{A digital object identifier}

\item{meta}{A list collecting all metadata used to create , Default: list()}
}
\value{
A Zotero-type matrix (tibble)
}
\description{
Connects with doi.org to create metadata
}
\seealso{
\code{\link[httr]{http_error}}, \code{\link[httr]{GET}},
 \code{\link[httr]{RETRY}}
 \code{\link[rvest]{rename}}, \code{\link[rvest]{html_attr}},
 \code{\link[rvest]{html_text}}, \code{\link[rvest]{reexports}},
 \code{\link[rvest]{html_children}}, \code{\link[rvest]{html_name}}
 \code{\link[dplyr]{bind}}, \code{\link[dplyr]{arrange}}
}
