% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero_format.R
\name{ZoteroFormat}
\alias{ZoteroFormat}
\title{Make a Zotero friendly format of input data}
\usage{
ZoteroFormat(
  data = NULL,
  format = NULL,
  prefix = NULL,
  check.structure = FALSE,
  fix.columns = TRUE,
  silent = FALSE
)
}
\arguments{
\item{data}{A list of metadata or something else, Default: NULL}

\item{format}{The format of the input data (e.g., 'JSON', 'versions', 'keys).
, Default: NULL}

\item{prefix}{Add a prefix to the metadata (e.g., user/userID), Default: NULL}

\item{check.structure}{Check that the structure of a data frame is correct
according to the Zotero type, Default: FALSE}

\item{fix.columns}{Remove columns not part of the Zotero type, Default: TRUE}

\item{silent}{c2z is noisy, tell it to be quiet, Default: TRUE}
}
\value{
A zotero friendly tibble if requested otherwise format the data
according to format.
}
\description{
Make a Zotero friendly format of input data, will add a zotero
key.
}
\examples{
\donttest{
  ZoteroFormat(
    list(title = "This is a test", itemType = "document")
  )
}
}
\seealso{
\code{\link[tibble]{tibble}}, \code{\link[tibble]{as_tibble}}
 \code{\link[jsonlite]{toJSON, fromJSON}}
 \code{\link[dplyr]{mutate_all}}, \code{\link[dplyr]{na_if}},
 \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{across}},
 \code{\link[dplyr]{reexports}}, \code{\link[dplyr]{select}}
 \code{\link[purrr]{map}}
}
