% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero_key.R
\name{ZoteroKey}
\alias{ZoteroKey}
\title{Create Zotero key or write token}
\usage{
ZoteroKey(token = FALSE)
}
\arguments{
\item{token}{Create a write token, Default: FALSE}
}
\value{
Will return a 8 character Zotero key or 32 character write token
}
\description{
Create an identifier for Zotero items/collections or a write
  token to library
}
\details{
Please see
  \href{https://oeysan.github.io/c2z/}{https://oeysan.github.io/c2z/}
}
\examples{
# Create a Zotero key
key <- ZoteroKey()
# Create a write token
token <- ZoteroKey(TRUE)
}
