% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero_types.R
\name{ZoteroTypes}
\alias{ZoteroTypes}
\title{List with empty zotero-items}
\format{
A list with 36 tibbles with zero rows and various columns
}
\usage{
ZoteroTypes(type = NULL, names = TRUE)
}
\arguments{
\item{type}{Type of zotero-item, Default: NULL}

\item{names}{return only column names if set to TRUE, Default: TRUE}
}
\value{
Either list of zotero-items or specified item
}
\description{
Each tibble in the list represents a zotero-item
}
\details{
Used to create Zotero-items from list of metadata
}
\examples{
# All zotero-items
names(ZoteroTypes())
# Column names of item-type `book`
ZoteroTypes("book")
}
