% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cristin_supported.R
\name{CristinSupported}
\alias{CristinSupported}
\title{Check for supported export formats from Cristin to Zotero}
\usage{
CristinSupported(data = NULL)
}
\arguments{
\item{data}{Zotero-type matrix, Default: NULL}
}
\value{
Zotero supported items with unsupported categories as NA
}
\description{
Used to filter Cristin categories that are not supported
}
\details{
Please see \href{https://oeysan.github.io/c2z/}{https://oeysan.github.io/c2z/}
}
\examples{
\dontrun{
  if(interactive()){
    # Supported Cristin items
    print(CristinSupported(), n = 32)
  }
}
}
\seealso{
\code{\link[tibble]{tibble}}
}
