% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cristin_wrangler.R
\name{CristinWrangler}
\alias{CristinWrangler}
\title{Wrangle Cristin metadata into Zotero-type format}
\usage{
CristinWrangler(
  data,
  zotero.check = TRUE,
  zotero = NULL,
  use.identifiers = TRUE,
  crossref.search = FALSE,
  autosearch = TRUE,
  remove.na = TRUE,
  replace.na = "book",
  force.type = NULL,
  override = FALSE,
  silent = FALSE,
  polite = TRUE,
  log = list()
)
}
\arguments{
\item{data}{Tibble containing metadata from Cristin}

\item{zotero.check}{Should the function look for duplicates?, Default: TRUE}

\item{zotero}{List with Zotero information used to check for duplicates,
Default: NULL}

\item{use.identifiers}{Use if ISBN/DOI identifiers if enabled, Default: TRUE}

\item{crossref.search}{Query Crossref database based on title, authors,
and date if enabled, Default: FALSE}

\item{autosearch}{Results could be automatically evaluated (based on some
logic) or you could inspect them manually if set to FALSE, Default: TRUE}

\item{remove.na}{Cristin contains many, more or less, obscure categories,
and not all are (yet) supported. By default these are removed, however, if
this option is set to FALSE unsupported categories are treated according to
replace.na, Default: TRUE}

\item{replace.na}{May the odds be in your favor and replace unsupported
categories with a predefined itemType if remove.na is set to
false, Default: 'book'}

\item{force.type}{Force all items to a predefined itemType, Default: NULL}

\item{override}{Put your faith in the algorithms and the identifiers
(i.e., DOI/ISNB) and override what is reported in Cristin, Default: FALSE}

\item{silent}{Running silent, running deep, Default: FALSE}

\item{polite}{Please store you email in `.Renviron` to query Crossref,
Default: TRUE}

\item{log}{A list for storing log elements, Default: list()}
}
\value{
A Zotero-type matrix (tibble)
}
\description{
The little engine that could, at least try, to wrangle metadata
from Cristin into a Zotero-type matrix. References can be augmented, or
replaced with metadata gathered through ISBN or DOI identifiers, and through
Crossref queries if all other fails. Or, hope the best and place your
misguided faith in the ability of researchers to correctly register their
own publications.
}
\seealso{
\code{\link[tidyr]{unnest}}, {\link[tidyr]{nest}}
 \code{\link[dplyr]{filter}}, \code{\link[dplyr]{across}},
 \code{\link[dplyr]{reexports}}, \code{\link[dplyr]{bind}},
 \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{distinct}}
 \code{\link[httr]{RETRY}}
 \code{\link[tibble]{tibble}}
}
