% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero_check.R
\name{ZoteroCheck}
\alias{ZoteroCheck}
\title{Check Zotero Library for Duplicates}
\usage{
ZoteroCheck(
  data,
  id,
  id.type,
  created,
  last.modified,
  items,
  remove.duplicates = TRUE,
  silent = FALSE,
  log = list()
)
}
\arguments{
\item{data}{A tibble containing metadata (e.g., from Cristin)}

\item{id}{A string specifying the column name in \code{data} that contains
the unique identifier (e.g., \code{"cristin_result_id"})}

\item{id.type}{A string specifying the type of metadata as stored in the
Zotero extra field (e.g., "Cristin")}

\item{created}{A string specifying the column name in \code{data} containing
the creation date in UNIX timestamp format (e.g., \code{"created"})}

\item{last.modified}{A string specifying the column name in \code{data}
containing the last modification date in UNIX timestamp format
(e.g., \code{"last_modified"})}

\item{items}{A data frame or tibble representing items in the Zotero library
to check against}

\item{remove.duplicates}{Logical. If \code{TRUE}, duplicates that have not
been modified are removed. Default is \code{TRUE}.}

\item{silent}{Logical. If \code{TRUE}, suppresses verbose output. Default
is \code{FALSE}.}

\item{log}{A list for storing log messages. Default is an empty list.}
}
\value{
A list with two elements: \code{data}, a tibble containing
non-duplicated items, and \code{log}, the updated log list.
}
\description{
Removes references from the input dataset that are duplicates
already added to Zotero,
preserving only those references that have been modified since their addition.
}
\details{
For more details,
see \href{https://oeysan.github.io/c2z/}{c2z documentation}.
}
\examples{
\dontrun{
  # Simple Cristin search by id
  cristin.data <- Cristin(
    id = "840998",
    zotero.import = FALSE
  )

  # Checking Zotero library for duplicates
  result <- ZoteroCheck(
    data = cristin.data$result,
    id = "cristin_result_id",
    id.type = "Cristin",
    created = "created",
    last.modified = "last_modified",
    items = Zotero(
      user = FALSE,
      id = "4827927",
      api = "Io1bwAiOjB1jPgUNegjJhJxF",
      library = TRUE,
      silent = TRUE
    )$items
  )
}
}
\seealso{
\code{\link[dplyr]{arrange}},
  \code{\link[dplyr]{coalesce}},
  \code{\link[dplyr]{bind_rows}},
  \code{\link[dplyr]{filter}}
}
